Add native active element feature
This feature allows to use the internal last active element instad of the
native active element which is available by the browser. This feature is
necessary for cases where you don't use the native focus system but need
access to the last element that was focues by this library. So, every time
when this library sets a focus it also saves the reference to that
element. That saved reference will be used to perform the blur event on
that element when the focus is changing. Also renamed the name of the
variable that saves the reference.